requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
scikit-learn>=1.3.0
joblib>=1.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
dnspython>=2.4.0